+++ integrations/zed/languages/cfxlua/injections.scm
@@ -1,0 +1,6 @@
+(((comment) @_luacats_comment
+  (#match? @_luacats_comment "^---")) @injection.content
+  (#set! injection.language "luacats"))
+
+((comment) @injection.content
+  (#set! injection.language "comment"))